Move fundingCreditHistory sync-schema to separate file
