+++ workers/loc.api/sync/schema/sync-schema/index.js
@@ -17,6 +17,7 @@ const orders = require('./orders')
 const movements = require('./movements')
 const fundingOfferHistory = require('./funding-offer-history')
 const fundingLoanHistory = require('./funding-loan-history')
+const fundingCreditHistory = require('./funding-credit-history')
 
 const _methodCollMap = new Map([
   [SYNC_API_METHODS.LEDGERS, ledgers],
@@ -28,21 +29,7 @@ const _methodCollMap = new Map([
   [SYNC_API_METHODS.MOVEMENTS, movements],
   [SYNC_API_METHODS.FUNDING_OFFER_HISTORY, fundingOfferHistory],
   [SYNC_API_METHODS.FUNDING_LOAN_HISTORY, fundingLoanHistory],
-  [
-    SYNC_API_METHODS.FUNDING_CREDIT_HISTORY,
-    {
-      name: ALLOWED_COLLS.FUNDING_CREDIT_HISTORY,
-      maxLimit: 10000,
-      dateFieldName: 'mtsUpdate',
-      symbolFieldName: 'symbol',
-      sort: [['mtsUpdate', -1]],
-      hasNewData: false,
-      start: [],
-      isSyncRequiredAtLeastOnce: false,
-      type: COLLS_TYPES.INSERTABLE_ARRAY_OBJECTS,
-      model: getModelOf(TABLES_NAMES.FUNDING_CREDIT_HISTORY)
-    }
-  ],
+  [SYNC_API_METHODS.FUNDING_CREDIT_HISTORY, fundingCreditHistory],
   [
     SYNC_API_METHODS.POSITIONS_HISTORY,
     {
